[NameFormatter] Move files from globals/ to file/ directory
