+++ src/concerts/ConcertController.cpp
@@ -6,10 +6,10 @@
 
 #include "concerts/Concert.h"
 #include "data/ImageCache.h"
+#include "file/NameFormatter.h"
 #include "globals/DownloadManager.h"
 #include "globals/Helper.h"
 #include "globals/Manager.h"
-#include "globals/NameFormatter.h"
 #include "media_centers/MediaCenterInterface.h"
 #include "scrapers/concert/ConcertScraperInterface.h"
 #include "settings/Settings.h"
