+++ src/file/NameFormatter.h
