+++ src/globals/ImageDialog.cpp
@@ -13,9 +13,9 @@
 #include <QtCore/qmath.h>
 
 #include "concerts/Concert.h"
+#include "file/NameFormatter.h"
 #include "globals/Helper.h"
 #include "globals/Manager.h"
-#include "globals/NameFormatter.h"
 #include "movies/Movie.h"
 #include "music/Album.h"
 #include "music/Artist.h"
