+++ src/ui/imports/ImportDialog.cpp
@@ -5,9 +5,9 @@
 #include <QMovie>
 
 #include "data/ImageCache.h"
+#include "file/NameFormatter.h"
 #include "globals/Helper.h"
 #include "globals/Manager.h"
-#include "globals/NameFormatter.h"
 #include "renamer/RenamerDialog.h"
 #include "scrapers/movie/CustomMovieScraper.h"
 #include "settings/Settings.h"
